AdapterService: Check if LeAudio device is capable to stream before making it active
Bug: 368997665
Test: mmm packages/modules/Bluetooth
Test: manual test
Flag: EXEMTP, obvious fix

+++ android/app/src/com/android/bluetooth/btservice/AdapterService.java
@@ -5074,11 +5074,15 @@ public class AdapterService extends Service {
                  */
                 mLeAudioService.removeActiveDevice(true /* hasFallbackDevice */);
             } else {
-                if (mA2dpService != null && mA2dpService.getActiveDevice() != null) {
-                    // TODO:  b/312396770
-                    mA2dpService.removeActiveDevice(false);
+                if (mLeAudioService.isGroupAvailableForStream(mLeAudioService.getGroupId(device))) {
+                    if (mA2dpService != null && mA2dpService.getActiveDevice() != null) {
+                        // TODO:  b/312396770
+                        mA2dpService.removeActiveDevice(false);
+                    }
+                    mLeAudioService.setActiveDevice(device);
+                } else {
+                    Log.w(TAG, "setActiveDevice: " + device + " not available for stream");
                 }
-                mLeAudioService.setActiveDevice(device);
             }
         }
